+++ github3/issues.py
@@ -122,16 +122,6 @@ class Milestone(GitHubCore):
         url = self._build_url('labels', base_url=self._api)
         return self._iter(int(number), url, Label)
 
-    def list_labels(self):
-        """List the labels for every issue associated with this
-        milestone.
-
-        :returns: list of :class:`Label <Label>`\ s
-        """
-        url = self._build_url('labels', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [Label(label, self) for label in json]
-
     @requires_auth
     def update(self, title, state='', description='', due_on=''):
         """Update this milestone.
@@ -339,15 +329,6 @@ class Issue(GitHubCore):
         url = self._build_url('comments', base_url=self._api)
         return self._iter(int(number), url, IssueComment)
 
-    def list_comments(self):
-        """List comments on this issue.
-
-        :returns: list of :class:`IssueComment <IssueComment>`
-        """
-        url = self._build_url('comments', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [IssueComment(comment, self) for comment in json]
-
     def iter_events(self, number=-1):
         """Iterate over events associated with this issue only.
 
@@ -357,15 +338,6 @@ class Issue(GitHubCore):
         """
         url = self._build_url('events', base_url=self._api)
         return self._iter(int(number), url, IssueEvent)
-
-    def list_events(self):
-        """List events associated with this issue only.
-
-        :returns: list of :class:`IssueEvent <IssueEvent>`\ s
-        """
-        url = self._build_url('events', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [IssueEvent(event, self) for event in json]
 
     @requires_auth
     def remove_label(self, name):