+++ github3/pulls.py
@@ -201,15 +201,6 @@ class PullRequest(GitHubCore):
         url = self._build_url('comments', base_url=self._api)
         return self._iter(int(number), url, ReviewComment)
 
-    def list_comments(self):
-        """List the comments on this pull request.
-
-        :returns: list of :class:`ReviewComment <ReviewComment>`\ s
-        """
-        url = self._build_url('comments', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [ReviewComment(comment, self) for comment in json]
-
     def iter_commits(self, number=-1):
         """Iterates over the commits on this pull request.
 
@@ -220,15 +211,6 @@ class PullRequest(GitHubCore):
         url = self._build_url('commits', base_url=self._api)
         return self._iter(int(number), url, Commit)
 
-    def list_commits(self):
-        """List the commits on this pull request.
-
-        :returns: list of :class:`Commit <github3.git.Commit>`\ s
-        """
-        url = self._build_url('commits', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [Commit(commit, self) for commit in json]
-
     def iter_files(self, number=-1):
         """Iterate over the files associated with this pull request.
 
@@ -238,15 +220,6 @@ class PullRequest(GitHubCore):
         """
         url = self._build_url('files', base_url=self._api)
         return self._iter(int(number), url, PullFile)
-
-    def list_files(self):
-        """List the files associated with this pull request.
-
-        :returns: list of :class:`PullFile <PullFile>`\ s
-        """
-        url = self._build_url('files', base_url=self._api)
-        json = self._json(self._get(url), 200)
-        return [PullFile(f) for f in json]
 
     @requires_auth
     def merge(self, commit_message=''):